# Conflicts:
#	src-client/mc/client/input/Keymapping.h
#	src/mc/dataloadhelper/NewUniqueIdsDataLoadHelper.h
#	src/mc/debug/PlayerRenderChunkModificationLatencyTracker.h
#	src/mc/deps/core/container/Blob.h
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/file/PathView.h
#	src/mc/deps/core/image/Image.h
#	src/mc/deps/core/math/Color.h
#	src/mc/deps/core/math/Vec2.h
#	src/mc/deps/core/math/Vec3.h
#	src/mc/deps/core/resource/PackIdVersion.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/deps/core/string/StringHash.h
#	src/mc/deps/core/utility/BinaryStream.h
#	src/mc/deps/core/utility/ReadOnlyBinaryStream.h
#	src/mc/deps/core_graphics/ImageDescription.h
#	src/mc/deps/ecs/WeakEntityRef.h
#	src/mc/deps/ecs/gamerefs_entity/EntityContext.h
#	src/mc/deps/ecs/gamerefs_entity/OwnerStorageEntity.h
#	src/mc/deps/ecs/gamerefs_entity/StackResultStorageEntity.h
#	src/mc/deps/ecs/gamerefs_entity/WeakStorageEntity.h
#	src/mc/deps/nbt/ByteArrayTag.h
#	src/mc/deps/nbt/ByteTag.h
#	src/mc/deps/nbt/CompoundTag.h
#	src/mc/deps/nbt/DoubleTag.h
#	src/mc/deps/nbt/FloatTag.h
#	src/mc/deps/nbt/Int64Tag.h
#	src/mc/deps/nbt/IntArrayTag.h
#	src/mc/deps/nbt/IntTag.h
#	src/mc/deps/nbt/ListTag.h
#	src/mc/deps/nbt/ShortTag.h
#	src/mc/deps/nbt/StringTag.h
#	src/mc/deps/raknet/RakNetGUID.h
#	src/mc/deps/raknet/SimpleMutex.h
#	src/mc/deps/raknet/SystemAddress.h
#	src/mc/entity/components/LookControlComponent.h
#	src/mc/entity/components_json_legacy/EconomyTradeableComponent.h
#	src/mc/external/webrtc/BufferT.h
#	src/mc/locale/I18nImpl.h
#	src/mc/network/NetworkIdentifier.h
#	src/mc/network/RakNetConnector.h
#	src/mc/network/packet/ModalFormRequestPacketPayload.h
#	src/mc/network/packet/ServerSettingsResponsePacketPayload.h
#	src/mc/network/packet/SpawnParticleEffectPacketPayload.h
#	src/mc/network/packet/StopSoundPacketPayload.h
#	src/mc/network/packet/TextPacketPayload.h
#	src/mc/network/packet/TransferPacket.h
#	src/mc/network/packet/TransferPacketPayload.h
#	src/mc/network/packet/UpdateAbilitiesPacket.h
#	src/mc/platform/CallStack.h
#	src/mc/platform/UUID.h
#	src/mc/platform/threading/PrioritizeSharedOwnership.h
#	src/mc/safety/RedactableString.h
#	src/mc/server/ServerPlayer.h
#	src/mc/server/SimulatedPlayer.h
#	src/mc/server/commands/Command.h
#	src/mc/server/commands/CommandBlockName.h
#	src/mc/server/commands/CommandMessage.h
#	src/mc/server/commands/CommandOrigin.h
#	src/mc/server/commands/CommandOutput.h
#	src/mc/server/commands/CommandParameterData.h
#	src/mc/server/commands/CommandRegistry.h
#	src/mc/server/commands/CommandVersion.h
#	src/mc/server/commands/ServerCommandOrigin.h
#	src/mc/server/module/RealmsStoriesGameplayEventListener.h
#	src/mc/util/ColorFormat.h
#	src/mc/util/MolangScriptArg.h
#	src/mc/util/MolangVariable.h
#	src/mc/util/MolangVariableMap.h
#	src/mc/util/VarIntDataInput.h
#	src/mc/util/VarIntDataOutput.h
#	src/mc/util/molang/ExpressionNode.h
#	src/mc/world/actor/Actor.h
#	src/mc/world/actor/ActorHurtResult.h
#	src/mc/world/actor/Mob.h
#	src/mc/world/actor/item/ItemActor.h
#	src/mc/world/actor/player/LayeredAbilities.h
#	src/mc/world/actor/player/Player.h
#	src/mc/world/actor/provider/SynchedActorDataAccess.h
#	src/mc/world/inventory/transaction/InventorySource.h
#	src/mc/world/item/ItemInstance.h
#	src/mc/world/item/ItemStack.h
#	src/mc/world/item/ItemStackBase.h
#	src/mc/world/item/components/OnUseItemComponent.h
#	src/mc/world/item/components/WeaponItemComponent.h
#	src/mc/world/level/BlockPos.h
#	src/mc/world/level/ChunkBlockPos.h
#	src/mc/world/level/ChunkPos.h
#	src/mc/world/level/SubChunkPos.h
#	src/mc/world/level/TickNextTickData.h
#	src/mc/world/level/biome/source/FixedBiomeSource.h
#	src/mc/world/level/block/Block.h
#	src/mc/world/level/block/actor/BlockActor.h
#	src/mc/world/level/block/registry/BlockTypeRegistry.h
#	src/mc/world/level/dimension/DimensionType.h
#	src/mc/world/level/dimension/VanillaDimensions.h
#	src/mc/world/level/levelgen/structure/BoundingBox.h
#	src/mc/world/level/levelgen/structure/PoolElementStructurePiece.h
#	src/mc/world/level/levelgen/structure/structurepools/StructurePoolBlockPredicateBlockMatchRandom.h
#	src/mc/world/phys/AABB.h
#	src/mc/world/phys/HitDetection.h
#	src/mc/world/phys/HitResult.h
#	src/mc/world/scores/PlayerScoreboardId.h
#	src/mc/world/scores/Scoreboard.h
#	src/mc/world/scores/ScoreboardId.h

+++ src/mc/world/level/block/NewWoodType.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class NewWoodType : int {
+enum class NewWoodType : uchar {
     Mangrove = 6,
     Bamboo   = 7,
     Cherry   = 8,
